Added the endpoint to fetch all accounts of a customer

+++ DomainModel/Account.cs
@@ -6,12 +6,12 @@ namespace DomainModel
     public class Account
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Type { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public string Currency { get; set; }
         public DateTime CreatedOn { get; set; }
         public DateTime ModifiedOn { get; set; }
-        public int CustomerId { get; set; }
+        public long CustomerId { get; set; }
     }
 }
